refactor(adapters): lift copilot-bridge into adapters/copilot/
Moves src/copilot-bridge.ts into src/adapters/copilot/adapter.ts wrapped as
CopilotSdkAttachment extends SdkAttachment extends BaseAttachment. Zero
behavior change — the entire main() body is now CopilotSdkAttachment.run(),
same lifecycle (createSession, event logger, workflow-ready wait, runId
pinning, processingStart/End wrapping around sendAndWait, cleanup).

The file stays dual-purpose: importable as a class (for the registry) AND
runnable directly via \`node .../adapter.js\`. The entry point is gated by
\`require.main === module\` so importing the class for its descriptor does
not trigger the bridge's main flow.

Registers copilotDescriptor in src/adapters/copilot/index.ts
(adapterId: 'copilot', class: 'sdk', blocksOnLLMTurn: true,
heartbeatMs: 30_000 per design §4.3).

src/spawn.ts resolveBridgePath() repointed from src/copilot-bridge.{ts,js}
to src/adapters/copilot/adapter.{ts,js}. Dev (ts-node) and prod (node) paths
both resolve through dist/adapters/copilot/ now.

Design reference: docs/design/session-lifecycle-rebuild-v2.md §6 (Class 2 —
SDK), §4.2, §4.4.
Sequencing memo: §3 PR-B.
